requirements.txt
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
mysqlclient==2.0.1
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
